nix: Bring back build flags lost in migration
These args were in the workflow but I deleted them in 91c916b736 and
forgot to bring them back.

Note that --baseURL is now hardcoded whereas before it was set from
hte output of the setup-pages GH Action. But there were already
hardcoded references to the base URL in the source code so whatever.